separate screenSwitch.
