+++ src/js/popup.js
@@ -2,6 +2,7 @@
 
 import { initDomainButton, setDomainEventListeners } from './domain.js';
 import { addHistory, setHistoryEventListeners } from './history.js';
+import { setScreenSwitchEventListeners } from './screenSwitch.js';
 import { initWhiteList, setWhiteListEventListeners } from './whitelist.js';
 
 const initKillOverWindow = () => {
@@ -18,6 +19,7 @@ const addEventListeners = () => {
   setWhiteListEventListeners();
   setDomainEventListeners();
   setHistoryEventListeners();
+  setScreenSwitchEventListeners();
 
   // checkbox event
   const checkElement = document.getElementById('target_all_windows');
@@ -64,30 +66,6 @@ const addEventListeners = () => {
     });
     addHistory(newHistoryFactors);
   });
-
-  // screen switch event
-  const screen_elements = document.getElementsByClassName('screen_switch');
-  for (const screen_element of screen_elements) {
-    screen_element.addEventListener('click', () =>
-      screenSwitcher(screen_element)
-    );
-  }
-};
-
-const screenSwitcher = (clicked_element) => {
-  const screen_elements = document.getElementsByClassName('screen_switch');
-  for (const screen_element of screen_elements) {
-    const screen_id = screen_element.id;
-    const parent = screen_element.parentElement;
-    const block_element = document.getElementById(screen_id + '_block');
-    if (clicked_element.id === screen_id) {
-      parent.classList.add('is-active');
-      block_element.style.display = 'block';
-    } else {
-      parent.classList.remove('is-active');
-      block_element.style.display = 'none';
-    }
-  }
 };
 
 initKillOverWindow();
